// tag-main/frontend_backup/src/api/apiTypes.ts
export interface RegisterPayload {
  name: string;
  email: string;
  userid: string;
  password: string;
  dob: string;
  gender: string;
  role: string;
}

export interface LoginPayload {
  email: string;
  password: string;
}

export interface RegisterResponse {
  id: string;
  name: string;
  userid: string;
  email: string;
  role: string;
  dob: string;
  gender: string;
  tokenExpiry?: string;
}

export interface LoginResponse {
  message: string;
  user: RegisterResponse;
}
// ...
export interface UpdatePasswordPayload {
  currentPassword: string;
  newPassword: string;
}

export interface UpdatePasswordResponse {
  message: string;
}
// ...
export interface RoleResponse {
  role: string;
}
// ...
export interface Guest {
  id: string;
  guest_name: string;
  ip_address: string;
  createdAt: string;
  updatedAt: string;
}

export interface User {
  id: string; // UUID
  userid: string; // username
  name: string;
  dob: string; // ISO date string (YYYY-MM-DD)
  gender: "male" | "female" | "other";
  email: string;
  password: string; // hashed password
  role: "user" | "admin" | "support";
  status: "active" | "inactive" | "locked";
  failed_login_attempts: number;
  last_seen: string; // ISO datetime string
  community_rating: number;
  createdAt: string; // ISO datetime
  updatedAt: string; // ISO datetime
}

export interface UserDetailsResponse {
  title: string;
  users: User[];
  pagination: paginate;
}

export interface paginate {
  total_items: number;
  current_page: number;
  totalPages: number;
  pageSize: number;
}
// ...
export interface UserSummaryStatsResponse {
  success: true;
  data: {
    total_registered_users: number;
    inactive_users: number;
    active_users: number;
    locked_users: number;
    support_users: number;
    admin_users: number;
    guest_users: number;
  };
}